fix: combine wallet + products
