+++ apps/web/app/settings/_menu.tsx
@@ -16,10 +16,6 @@ const menuItems: Record<
     icon: <Icon.Wallet size="24" />,
     label: 'Wallet',
   },
-  products_services: {
-    icon: <Icon.Product size="24" />,
-    label: 'Products & Services',
-  },
   privacy_safety: {
     icon: <Icon.Shield size="24" />,
     label: 'Privacy & Safety',
