+++ apps/web/app/settings/sections/index.tsx
@@ -4,7 +4,6 @@ import Language from './_language';
 import MutedUsers from './_muted';
 import Notifications from './_notifications';
 import PrivacySafety from './_privacy';
-import ProductsServices from './_products_services';
 import Wallet from './_wallet';
 
 export const Section = {
@@ -14,6 +13,5 @@ export const Section = {
   MutedUsers,
   Notifications,
   PrivacySafety,
-  ProductsServices,
   Wallet,
 };
